Refactor zipper implementation
This implementation makes it much easier to implement the setter methods
on the zipper. The idea here is that the trees are stored in a stack as
moves are made. Updates occur on the current tree without affecting the
others. The key is that when the zipper is moved up, the parent tree is
then mutated to receive the changes in the tree. I can only get away
with this using Elixir's immutable data structures. 🙂